+++ src/test/groovy/biz/nellemann/hmci/LogicalPartitionTest.groovy
@@ -139,4 +139,31 @@ class LogicalPartitionTest extends Specification {
         listOfMeasurements.first().fields['readBytes'] == 0.0
     }
 
+    void "test getSriovLogicalPortMetrics'"() {
+
+        setup:
+        def testFile = new File(getClass().getResource('/pcm-data-logical-partition-sriov.json').toURI())
+        def testJson = testFile.getText('UTF-8')
+        ManagedSystem system = new ManagedSystem("e09834d1-c930-3883-bdad-405d8e26e166", "Test Name","Test Type", "Test Model", "Test S/N")
+        LogicalPartition lpar = new LogicalPartition("2DE05DB6-8AD5-448F-8327-0F488D287E82", "9Flash01", "OS400", system)
+
+        when:
+        lpar.processMetrics(testJson)
+        List<Measurement> listOfMeasurements = lpar.getSriovLogicalPorts()
+
+        then:
+        listOfMeasurements.size() == 6
+        listOfMeasurements.first().tags['location'] == "U78CA.001.CSS0CXA-P1-C2-C1-T1-S2"
+        listOfMeasurements.first().tags['vnicDeviceMode'] == "NonVNIC"
+        listOfMeasurements.first().tags['configurationType'] == "Ethernet"
+        listOfMeasurements.first().fields['drcIndex'] == "654327810"
+        listOfMeasurements.first().fields['physicalPortId'] == 0
+        listOfMeasurements.first().fields['physicalDrcIndex'] == "553713681"
+        listOfMeasurements.first().fields['receivedPackets'] == 16.867
+        listOfMeasurements.first().fields['sentPackets'] == 0.067
+        listOfMeasurements.first().fields['sentBytes'] == 8.533
+        listOfMeasurements.first().fields['receivedBytes'] == 1032.933
+        listOfMeasurements.first().fields['transferredBytes'] == 1041.466
+    }
+
 }